refactor: NicovideoRegistrationRequestのディレクトリ整理
Fixes #612
