+++ src/NicovideoRegistrationRequest/NicovideoRegistrationRequestRejecting.resolver.ts
@@ -1,7 +1,7 @@
-import { Resolvers } from "../graphql.js";
-import { GraphQLNotExistsInDBError } from "../id.js";
-import { NicovideoRegistrationRequestModel } from "../NicovideoRegistrationRequest/model.js";
-import { ResolverDeps } from "../types.js";
+import { Resolvers } from "../resolvers/graphql.js";
+import { GraphQLNotExistsInDBError } from "../resolvers/id.js";
+import { ResolverDeps } from "../resolvers/types.js";
+import { NicovideoRegistrationRequestDTO } from "./dto.js";
 
 export const resolverNicovideoRegistrationRequestRejecting = ({
   prisma,
@@ -11,7 +11,7 @@ export const resolverNicovideoRegistrationRequestRejecting = ({
     request: ({ requestId }) =>
       prisma.nicovideoRegistrationRequest
         .findUniqueOrThrow({ where: { id: requestId } })
-        .then((u) => new NicovideoRegistrationRequestModel(u))
+        .then((u) => new NicovideoRegistrationRequestDTO(u))
         .catch(() => {
           throw new GraphQLNotExistsInDBError("NicovideoRegistrationRequest", requestId);
         }),
