+++ src/NicovideoRegistrationRequest/findNicovideoRegistrationRequests.resolver.ts
@@ -2,12 +2,12 @@ import { findManyCursorConnection } from "@devoxa/prisma-relay-cursor-connection
 import { GraphQLError } from "graphql";
 import z from "zod";
 
-import { isErr } from "../../../utils/Result.js";
-import { cursorOptions } from "../../connection.js";
-import { QueryResolvers } from "../../graphql.js";
-import { NicovideoRegistrationRequestConnectionModel } from "../../NicovideoRegistrationRequestConnection/model.js";
-import { parseOrderBy } from "../../parseSortOrder.js";
-import { ResolverDeps } from "../../types.js";
+import { cursorOptions } from "../resolvers/connection.js";
+import { QueryResolvers } from "../resolvers/graphql.js";
+import { parseOrderBy } from "../resolvers/parseSortOrder.js";
+import { ResolverDeps } from "../resolvers/types.js";
+import { isErr } from "../utils/Result.js";
+import { NicovideoRegistrationRequestConnectionDTO } from "./dto.js";
 
 export const findNicovideoRegistrationRequests = ({ prisma, logger }: Pick<ResolverDeps, "prisma" | "logger">) =>
   (async (_, { orderBy: unparsedOrderBy, checked, ...unparsedConnectionArgs }, { currentUser: ctxUser }, info) => {
@@ -47,5 +47,5 @@ export const findNicovideoRegistrationRequests = ({ prisma, logger }: Pick<Resol
       () => prisma.nicovideoRegistrationRequest.count({ where: { isChecked: checked?.valueOf() } }),
       connectionArgs.data,
       { resolveInfo: info, ...cursorOptions }
-    ).then((c) => NicovideoRegistrationRequestConnectionModel.fromPrisma(c));
+    ).then((c) => NicovideoRegistrationRequestConnectionDTO.fromPrisma(c));
   }) satisfies QueryResolvers["findNicovideoRegistrationRequests"];
