+++ src/NicovideoRegistrationRequest/getNicovideoRegistrationRequest.ts
@@ -1,13 +1,13 @@
-import { QueryResolvers } from "../../graphql.js";
-import { GraphQLNotExistsInDBError, parseGqlID } from "../../id.js";
-import { NicovideoRegistrationRequestModel } from "../../NicovideoRegistrationRequest/model.js";
-import { ResolverDeps } from "../../types.js";
+import { QueryResolvers } from "../resolvers/graphql.js";
+import { GraphQLNotExistsInDBError, parseGqlID } from "../resolvers/id.js";
+import { ResolverDeps } from "../resolvers/types.js";
+import { NicovideoRegistrationRequestDTO } from "./dto.js";
 
 export const getNicovideoRegistrationRequest = ({ prisma, logger }: Pick<ResolverDeps, "prisma" | "logger">) =>
   (async (_parent, { id }, { currentUser: ctxUser }, info) =>
     prisma.nicovideoRegistrationRequest
       .findUniqueOrThrow({ where: { id: parseGqlID("NicovideoRegistrationRequest", id) } })
-      .then((v) => new NicovideoRegistrationRequestModel(v))
+      .then((v) => new NicovideoRegistrationRequestDTO(v))
       .catch(() => {
         logger.error({ path: info.path, args: { id }, userId: ctxUser?.id }, "Not found");
         throw new GraphQLNotExistsInDBError("NicovideoRegistrationRequest", id);
