--- src/resolvers/Mutation/requestNicovideoRegistration/requestNicovideoRegistration.graphql
@@ -1,39 +1,0 @@
-type Mutation {
-  requestNicovideoRegistration(input: RequestNicovideoRegistrationInput!): RequestNicovideoRegistrationReturnUnion!
-    @auth(scopes: ["create:registration_request"])
-}
-
-input RequestNicovideoRegistrationInput {
-  sourceId: String!
-
-  title: String!
-  thumbnailUrl: String!
-
-  taggings: [RequestNicovideoRegistrationInputTagging!]!
-  semitaggings: [RequestNicovideoRegistrationInputSemitagging!]!
-}
-
-input RequestNicovideoRegistrationInputTagging {
-  tagId: ID!
-  note: String
-}
-
-input RequestNicovideoRegistrationInputSemitagging {
-  name: String!
-  note: String
-}
-
-union RequestNicovideoRegistrationReturnUnion =
-    MutationInvalidTagIdError
-  | MutationTagNotFoundError
-  | RequestNicovideoRegistrationVideoAlreadyRegisteredError
-  | RequestNicovideoRegistrationSucceededPayload
-
-"動画が既に登録されている場合のエラー"
-type RequestNicovideoRegistrationVideoAlreadyRegisteredError {
-  source: NicovideoVideoSource!
-}
-
-type RequestNicovideoRegistrationSucceededPayload {
-  request: NicovideoRegistrationRequest!
-}
