--- src/resolvers/NicovideoRegistrationRequest/NicovideoRegistrationRequest.graphql
@@ -1,36 +1,0 @@
-type NicovideoRegistrationRequest {
-  id: ID!
-  title: String!
-
-  requestedBy: User!
-
-  sourceId: String!
-
-  title: String!
-  thumbnailUrl: String!
-
-  taggings: [NicovideoRegistrationRequestTagging!]!
-  semitaggings: [NicovideoRegistrationRequestSemitagging!]!
-
-  checked: Boolean!
-  checking: NicovideoRegistrationRequestChecking
-
-  originalUrl: String!
-  embedUrl: String!
-}
-
-type NicovideoRegistrationRequestTagging {
-  id: ID!
-  tag: Tag!
-  note: String
-}
-
-type NicovideoRegistrationRequestSemitagging {
-  id: ID!
-  name: String!
-  note: String
-}
-
-union NicovideoRegistrationRequestChecking =
-    NicovideoRegistrationRequestAccepting
-  | NicovideoRegistrationRequestRejecting
