--- src/resolvers/NicovideoRegistrationRequestAcceptingNotification/NicovideoRegistrationRequestAcceptingNotification.graphql
@@ -1,8 +1,0 @@
-type NicovideoRegistrationRequestAcceptingNotification implements Node & Notification {
-  id: ID!
-  createdAt: DateTime!
-  notifyTo: User!
-  watched: Boolean!
-
-  accepting: NicovideoRegistrationRequestAccepting!
-}
