--- src/resolvers/NicovideoRegistrationRequestConnection/NicovideoRegistrationRequest.graphql
@@ -1,11 +1,0 @@
-type NicovideoRegistrationRequestConnection {
-  nodes: [NicovideoRegistrationRequest!]!
-  edges: [NicovideoRegistrationRequestEdge!]!
-  pageInfo: PageInfo!
-  totalCount: Int!
-}
-
-type NicovideoRegistrationRequestEdge {
-  cursor: String!
-  node: NicovideoRegistrationRequest!
-}
