--- src/resolvers/NicovideoRegistrationRequestRejecting/NicovideoRegistrationRequestRejecting.graphql
@@ -1,7 +1,0 @@
-type NicovideoRegistrationRequestRejecting {
-  request: NicovideoRegistrationRequest!
-
-  note: String
-
-  rejectedBy: User!
-}
