--- src/resolvers/NicovideoRegistrationRequestRejectingNotification/NicovideoRegistrationRequestRejectingNotification.graphql
@@ -1,8 +1,0 @@
-type NicovideoRegistrationRequestRejectingNotification implements Node & Notification {
-  id: ID!
-  createdAt: DateTime!
-  notifyTo: User!
-  watched: Boolean!
-
-  rejecting: NicovideoRegistrationRequestRejecting!
-}
