--- src/resolvers/Query/findNicovideoRegistrationRequest/findNicovideoRegistrationRequest.graphql
@@ -1,8 +1,0 @@
-type Query {
-  findNicovideoRegistrationRequest(input: FindNicovideoRegistrationRequestInput): NicovideoRegistrationRequest
-}
-
-input FindNicovideoRegistrationRequestInput {
-  id: ID
-  sourceId: String
-}
