--- src/resolvers/Query/findNicovideoRegistrationRequests/findNicovideoRegistrationRequests.graphql
@@ -1,17 +1,0 @@
-type Query {
-  findNicovideoRegistrationRequests(
-    first: Int
-    after: String
-
-    last: Int
-    before: String
-
-    checked: Boolean
-
-    orderBy: FindNicovideoRegistrationRequestsOrderBy! = { createdAt: DESC }
-  ): NicovideoRegistrationRequestConnection!
-}
-
-input FindNicovideoRegistrationRequestsOrderBy {
-  createdAt: SortOrder
-}
