--- src/resolvers/Query/findUncheckedNicovideoRegistrationRequests/findUncheckedNicovideoRegistrationRequests.graphql
@@ -1,19 +1,0 @@
-type Query {
-  findUncheckedNicovideoRegistrationRequests(
-    input: FindUncheckedNicovideoRegistrationRequestsInput!
-  ): NicovideoRegistrationRequestConnection!
-}
-
-input FindUncheckedNicovideoRegistrationRequestsInput {
-  first: Int
-  after: String
-
-  last: Int
-  before: String
-
-  orderBy: FindUncheckedNicovideoRegistrationRequestsOrderBy! = { createdAt: DESC }
-}
-
-input FindUncheckedNicovideoRegistrationRequestsOrderBy {
-  createdAt: SortOrder
-}
