--- src/resolvers/Query/getNicovideoRegistrationRequest/getNicovideoRegistrationRequest.graphql
@@ -1,3 +1,0 @@
-type Query {
-  getNicovideoRegistrationRequest(id: ID!): NicovideoRegistrationRequest!
-}
